+++ derive/__init__.py
@@ -1,0 +1,6 @@
+"""
+Init for the derive client
+"""
+from .derive import DeriveClient
+
+DeriveClient
