--- lyra/__init__.py
@@ -1,6 +1,0 @@
-"""
-Init for the lyra client
-"""
-from .lyra import LyraClient
-
-LyraClient
